requirements.txt
Flask==3.0.3
rich==13.8.1
requests==2.32.3
ipinfo==5.1.1
python-dotenv==1.0.1
geopy==2.4.1